rename src/components/FAQHelp/FAQHelp.tsx to index.tsx
